+++ src/main/java/model/Request/RequestLine.java
@@ -1,0 +1,31 @@
+package model.Request;
+
+public class RequestLine {
+    private final String version;
+    private final String method;
+    private final String url;
+
+    public RequestLine(String reqLine) {
+        String[] tokens = reqLine.split(" ");
+        this.method = tokens[0];
+
+        String url = tokens[1];
+        if (url.equals("/")) {
+            url = "/index.html";
+        }
+        this.url = url;
+        this.version = tokens[2];
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public String getMethod() {
+        return method;
+    }
+
+    public String getUrl() {
+        return url;
+    }
+}
