+++ src/test/java/service/UserServiceTest.java
@@ -2,10 +2,9 @@ package service;
 
 import db.Database;
 import exception.DuplicateUserIdException;
-import model.Request;
+import model.Request.Request;
 import model.User;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import webserver.UserService;
